Small Gamma: Plotting D
This new function plots D-stat on the y-axis against the size of the
first sample (m) on the x-axis. It create the graph for three values of
gamma (under 0.5).

So far, it can simulate from mv normal and mv t distribution. It would
be easy to add more distribution with if statements.

The goal: plot nice graphs and eye ball them for analysis.